Move `move_accessor.py` to the top level
