--- movement/io/__init__.py
@@ -1,1 +1,0 @@
-from .move_accessor import MoveAccessor
